services/vision-backend/requirements.txt
# --- Core server ---
flask==3.1.2
flask-cors==5.0.0
python-dotenv==1.0.1

# --- Vision / AI ---
torch==2.8.0
torchvision==0.23.0
onnxruntime-gpu==1.23.2
ultralytics==8.3.34
opencv-python-headless==4.10.0.84
face-recognition==1.3.0

# --- Speech & TTS ---
SpeechRecognition==3.10.4
pyttsx3==2.91

# --- Utilities ---
Pillow==11.0.0
numpy==2.2.1
google-generativeai==0.7.2
